Update the Permissions html. And convert the web view fragments to Jetpack Compose
