+++ app/src/main/assets/files/style.css
@@ -1,0 +1,100 @@
+:root {
+  color-scheme: light dark;
+  --bg: #f5f5f7;
+  --card-bg: #ffffff;
+  --text: #1c1c1e;
+  --muted: #6e6e73;
+  --border: #e2e2e5;
+  --accent: #3366cc;
+  --badge-bg: #eef1fb;
+  --badge-text: #3355aa;
+}
+@media (prefers-color-scheme: dark) {
+  :root {
+    --bg: #121212;
+    --card-bg: #1e1e1e;
+    --text: #f0f0f2;
+    --muted: #a3a3a8;
+    --border: #333336;
+    --accent: #8ab4ff;
+    --badge-bg: #26314f;
+    --badge-text: #a8c2ff;
+  }
+}
+* { box-sizing: border-box; }
+body {
+  margin: 0;
+  padding: 24px 16px 48px;
+  background: var(--bg);
+  color: var(--text);
+  font-family: -apple-system, "Segoe UI", Roboto, Helvetica, Arial, sans-serif;
+  line-height: 1.5;
+}
+.container { max-width: 720px; margin: 0 auto; }
+h1 {
+  font-size: 20px;
+  font-weight: 600;
+  margin: 0 0 4px;
+}
+.subtitle {
+  color: var(--muted);
+  font-size: 14px;
+  margin: 0 0 24px;
+}
+details {
+  background: var(--card-bg);
+  border: 1px solid var(--border);
+  border-radius: 12px;
+  margin-bottom: 10px;
+  overflow: hidden;
+}
+summary {
+  list-style: none;
+  cursor: pointer;
+  padding: 14px 16px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+  font-weight: 600;
+  font-size: 15px;
+}
+summary::-webkit-details-marker { display: none; }
+summary::after {
+  content: "\25BE";
+  color: var(--muted);
+  font-weight: 400;
+  transition: transform 0.2s ease;
+  flex-shrink: 0;
+}
+details[open] summary::after { transform: rotate(180deg); }
+.summary-title { flex: 1; }
+.badge {
+  font-size: 11px;
+  font-weight: 600;
+  color: var(--badge-text);
+  background: var(--badge-bg);
+  padding: 3px 9px;
+  border-radius: 999px;
+  white-space: nowrap;
+}
+.section-body {
+  padding: 0 16px 16px;
+  border-top: 1px solid var(--border);
+}
+.section-body p {
+  color: var(--muted);
+  font-size: 13px;
+  margin: 12px 0 0;
+}
+.copyright {
+  color: var(--text) !important;
+  font-weight: 500;
+}
+footer {
+  text-align: center;
+  color: var(--muted);
+  font-size: 12px;
+  margin-top: 20px;
+}
+a { color: var(--accent); }
