+++ app/src/main/java/com/elementary/tasks/settings/other/PermissionsFragment.kt
@@ -1,28 +1,11 @@
 package com.elementary.tasks.settings.other
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import com.elementary.tasks.R
-import com.elementary.tasks.databinding.FragmentSettingsWebViewBinding
-import com.elementary.tasks.navigation.fragments.BaseSettingsFragment
+import com.elementary.tasks.navigation.toolbarfragment.BaseComposeWebViewFragment
 
-class PermissionsFragment : BaseSettingsFragment<FragmentSettingsWebViewBinding>() {
-  override fun inflate(
-    inflater: LayoutInflater,
-    container: ViewGroup?,
-    savedInstanceState: Bundle?,
-  ) = FragmentSettingsWebViewBinding.inflate(inflater, container, false)
-
-  override fun onViewCreated(
-    view: View,
-    savedInstanceState: Bundle?,
-  ) {
-    super.onViewCreated(view, savedInstanceState)
-    val url = "file:///android_asset/files/permissions.html"
-    binding.webView.loadUrl(url)
-  }
+class PermissionsFragment : BaseComposeWebViewFragment() {
+  override val url: String
+    get() = "file:///android_asset/files/permissions.html"
 
   override fun getTitle(): String = getString(R.string.permissions)
 }
