+++ app/src/main/java/com/elementary/tasks/settings/other/PrivacyPolicyFragment.kt
@@ -1,45 +1,14 @@
 package com.elementary.tasks.settings.other
 
-import android.annotation.SuppressLint
-import android.webkit.WebChromeClient
-import android.webkit.WebView
-import android.webkit.WebViewClient
 import com.elementary.tasks.R
-import com.elementary.tasks.navigation.fragments.BaseWebViewFragment
+import com.elementary.tasks.navigation.toolbarfragment.BaseComposeWebViewFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class PrivacyPolicyFragment : BaseWebViewFragment() {
+class PrivacyPolicyFragment : BaseComposeWebViewFragment() {
   private val viewModel by viewModel<PrivacyPolicyViewModel>()
 
   override val url: String
     get() = viewModel.url
-
-  @SuppressLint("SetJavaScriptEnabled")
-  override fun setExtraParams(webView: WebView) {
-    super.setExtraParams(webView)
-    webView.settings.javaScriptEnabled = true
-    webView.webViewClient =
-      object : WebViewClient() {
-        @Deprecated("Deprecated in Java")
-        override fun onReceivedError(
-          view: WebView,
-          errorCode: Int,
-          description: String,
-          failingUrl: String,
-        ) {
-        }
-
-        @Deprecated("Deprecated in Java")
-        override fun shouldOverrideUrlLoading(
-          view: WebView,
-          url: String,
-        ): Boolean {
-          view.loadUrl(url)
-          return true
-        }
-      }
-    webView.webChromeClient = WebChromeClient()
-  }
 
   override fun getTitle(): String = getString(R.string.privacy_policy)
 }
