--- app/src/main/res/layout/fragment_settings_whats_new.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.core.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="?android:colorBackground"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="24dp"
-        android:layout_marginEnd="24dp"
-        android:orientation="vertical">
-
-        <Space
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/collapse_toolbar_margin_top" />
-
-        <TextView
-            android:id="@+id/dateAndVersionView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/whats_new_version_and_date"
-            android:textAppearance="?textAppearanceHeadlineSmall"
-            android:textColor="?colorTertiary" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="16dp"
-            android:text="@string/whats_new_text"
-            android:textAppearance="?textAppearanceTitleMedium" />
-
-    </LinearLayout>
-
-</androidx.core.widget.NestedScrollView>
